feat(flight): Phase 1 foundation — arrow feature split, cqlite-flight crate, ticket module
Split CQL→Arrow conversion out of the parquet writer into a new
cqlite-core 'arrow' feature (export::arrow_convert) so the Flight server
can reuse it without depending on the parquet crate. parquet now builds
on arrow. Existing parquet tests pass unchanged.

Add the cqlite-flight crate (workspace member) and its ticket module:
FlightTicket is pure-serde JSON (keyspace, table, ddl, snapshot, token
range, projection, predicates) carrying token_in_range() with Cassandra
(start, end] + wraparound semantics. 8 TDD unit tests.

Plan and append-only journal under docs/flight-trino/.

Refs: docs/flight-trino/PLAN.md (Phase 1)
